requirements.txt
django
djangorestframework
django-allauth
django-rest-auth-dj4
django-rest-passwordreset
drf-yasg
django-cors-headers
Pillow
psycopg2-binary
python-dotenv
pandas
openpyxl